05-16 리뷰 수정 경환 commit

+++ src/main/java/hannyanggang/catchdoctor/dto/ReviewDto.java
@@ -23,7 +23,7 @@ public class ReviewDto {
 
     private String username;
 
-    private int grade;
+    private float grade;
 
     private String text;
